Move tests and test util function as well.
